backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0

google-analytics-data==0.18.7
google-analytics-admin==0.20.0
google-api-core==2.25.1
google-auth==2.38.0
google-auth-oauthlib==1.2.1

grpcio==1.62.3
grpcio-status==1.62.3
protobuf==4.25.3

pandas==2.2.3
numpy==1.26.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
anthropic
